+++ packages/lib/src/typeChecking/object.ts
@@ -4,7 +4,14 @@ import { assertIsFunction, assertIsObject, isObject, lateVal } from "../utils"
 import { getTypeInfo } from "./getTypeInfo"
 import { resolveStandardType, resolveTypeChecker } from "./resolveTypeChecker"
 import type { AnyStandardType, AnyType, ObjectType, ObjectTypeFunction } from "./schemas"
-import { lateTypeChecker, LateTypeChecker, TypeChecker, TypeInfo, TypeInfoGen } from "./TypeChecker"
+import {
+  lateTypeChecker,
+  LateTypeChecker,
+  TypeChecker,
+  TypeCheckerBaseType,
+  TypeInfo,
+  TypeInfoGen,
+} from "./TypeChecker"
 import { TypeCheckError } from "./TypeCheckError"
 
 function typesObjectHelper<S>(objFn: S, frozen: boolean, typeInfoGen: TypeInfoGen): S {
@@ -33,16 +40,19 @@ function typesObjectHelper<S>(objFn: S, frozen: boolean, typeInfoGen: TypeInfoGe
     }
 
     const thisTc: TypeChecker = new TypeChecker(
+      TypeCheckerBaseType.Object,
+
       (obj, path) => {
-        if (!isObject(obj) || (frozen && !(obj instanceof Frozen)))
+        if (!isObject(obj) || (frozen && !(obj instanceof Frozen))) {
           return new TypeCheckError(path, getTypeName(thisTc), obj)
+        }
 
         // note: we allow excess properties when checking objects
         for (const [k, unresolvedTc] of schemaEntries) {
           const tc = resolveTypeChecker(unresolvedTc)
           const objVal = obj[k]
 
-          const valueError = !tc.unchecked ? tc.check(objVal, [...path, k]) : null
+          const valueError = tc.check(objVal, [...path, k])
           if (valueError) {
             return valueError
           }
@@ -50,8 +60,48 @@ function typesObjectHelper<S>(objFn: S, frozen: boolean, typeInfoGen: TypeInfoGe
 
         return null
       },
+
       getTypeName,
-      typeInfoGen
+      typeInfoGen,
+
+      (obj) => {
+        if (!isObject(obj)) {
+          return null
+        }
+
+        // note: we allow excess properties when checking objects
+        for (const [k, unresolvedTc] of schemaEntries) {
+          const tc = resolveTypeChecker(unresolvedTc)
+          const objVal = obj[k]
+
+          const valueActualChecker = tc.snapshotType(objVal)
+          if (!valueActualChecker) {
+            return null
+          }
+        }
+
+        return thisTc
+      },
+
+      (obj: Record<string, unknown>) => {
+        const newObj: typeof obj = {}
+
+        // note: we allow excess properties when checking objects
+        const keys = Object.keys(obj)
+        for (let i = 0; i < keys.length; i++) {
+          const k = keys[i]
+          const unresolvedTc = objectSchema[k]
+          if (unresolvedTc) {
+            const tc = resolveTypeChecker(unresolvedTc)
+            newObj[k] = tc.fromSnapshotProcessor(obj[k])
+          } else {
+            // unknown prop, copy as is
+            newObj[k] = obj[k]
+          }
+        }
+
+        return newObj
+      }
     )
 
     return thisTc